requirements.txt
numpy==1.26.4
pandas==2.2.3
matplotlib>=3.4.0
pyqt5>=5.15.0
pyqtgraph>=0.12.0
lightgbm==4.5.0
scikit-learn==1.5.2
joblib>=1.4.2
PyYAML>=5.4.0
